requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.15.4
filterpy==1.4.5
fonttools==4.53.1
fsspec==2024.6.1
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
mysql-connector-python==9.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pyparsing==3.1.2
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.5
six==1.16.0
sympy==1.13.1
tifffile==2024.7.24
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.73
ultralytics-thop==2.0.0
urllib3==2.2.2
zipp==3.19.2
